+++ CHANGELOG.md
@@ -6,6 +6,11 @@ Deprecated
 
 - :func:`sheraf.models.BaseIndexedAttribute.make_id`. Please use the 'default' parameter of your id :class:`~sheraf.attributes.base.BaseAttribute` instead.
 
+Added
+*****
+
+- :class:`sheraf.attributes.collections.SmallDictAttribute`.
+
 [0.1.0] - 2020-02-11
 ====================
 - First version.